commit 0.43 Updated transaction deposite

+++ src/main/java/org/example/kihelp_back/wallet/dto/WalletResponseDto.java
@@ -1,8 +1,10 @@
 package org.example.kihelp_back.wallet.dto;
 
+import java.math.BigDecimal;
+
 public record WalletResponseDto(
         String name,
-        Double balance,
+        BigDecimal balance,
         boolean isDefault
 ) {
 }
